Disable line numbering for listings that will be copy/pasted.
The line numbering for the listings that are expected to be copy/pasted
make things awkward to interact with and make the installation / getting
started a bit frustrating.

This is especially true for multi line commands which are common in the
getting started section.

By disabling the line numbering, copy/pasting becomes significantly
easier.